+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v1.1.0 (2022-04-12)
 - Changed bundler to rollup; better module generation, only generates browser module
 - Disabled behaviour; setting the element as disabled now works as expected
+- New method "clear"
 
 ### v1.0.0 (2021-05-24)
 - First release